requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pyperclip>=1.8.2
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0
winotify>=1.1.0
pyautogui>=0.9.54
